Bug 480683 - Icon Renderer: Folder "dgm" and big icons
The eclipse user interface guidelines says."The first letter of Toolbar
and Local toolbar folder names indicates the icon state. Use the letter
d for disabled, or e for enabled." Because of this the icons in e.g.
"dlcl16" folder are not handled by the icon renderer as they are derived
from the version in the "elcl16" folder. But there is one exception from
this rule: The "dgm" folder it starts with a "d" but does not contain
"disabled state" versions of icons. So the "dgm" folder must also be
processed by the icon renderer.

The gif compare gallery renderer assumed that the icons are 16x16 pixels
at max. In the other galleries icons are scaled down / up to 16x16
pixels. Icons that are bigger then 16x16 pixels messed up the gif
compare gallery. The compare gallery now scales down icons that higher /
wider then 16 pixels.

Change-Id: I6ca84701e73eaf472bc527afc80387f9278a9fbf
Signed-off-by: Matthias Becker <ma.becker@sap.com>