Create SVG variants for shared DnD icons
This creates SVG files for all images that can be accessed via
ISharedImages.IMG_OBJS_DND_*, as well as contributes the 100% and 200%
zoom png files that have been created from them.

In addition to the already existing files, this also contributes a new
variant where the source and mask are merged into a single image.

In order to get the best results when creating the source and mask
images, one should disable anti-aliasing.