Align parent group id
Currently the parent uses eclipse.images as its group id, while the
childs use ord.eclipse.images.

This aligns the parent group id with the childs making it redundant to
specify a groupid there again and making it more consistent, also
removing the relative path as it is the default.