Adapt readme with information about this repository being archived
The SVG files inside this repository have been embedded into the Eclipse
platform and and other bundles and their rasterized versions are not
required anymore due to runtime support for SVGs added to Eclipse SWT.

To avoid maintenance of icon sources at multiple places, this repository
is supposed to be archived. In order to document the archiving and the
places to find and browse the icons now, this change adapts the readme
accordingly.

Contributes to https://github.com/eclipse-platform/eclipse.platform.images/issues/152